// src/app/model/ResumeFormData.ts
export interface ResumeFormData {
    firstName: string;
    middleName: string;
    lastName: string;
    email: string;
    phone: string;
    desc: string;
    decl: string;
    web: string;
    university11: string;
    yearofpassing11: string;
    percentage11: string;
    university21: string;
    yearofpassing21: string;
    percentage21: string;
    university31: string;
    yearofpassing31: string;
    percentage31: string;
    university41: string;
    yearofpassing41: string;
    percentage41: string;
    designation1: string;
    organization1: string;
    doj1: string;
    dor1: string;
    designation2: string;
    organization2: string;
    doj2: string;
    dor2: string;
    designation3: string;
    organization3: string;
    doj3: string;
    dor3: string;
    designation4: string;
    organization4: string;
    doj4: string;
    dor4: string;


}